[ovirttool] validate clusters; installer arg
Two changes:
1. if cluster created, then once entire job has completed,
   will validate the cluster from each node, and print
   summary to user based on this gathered data.
   (currently, end of job summary relying on static data saved
   throughout job processing)

2. Require --installer option to be a valid URL that can use
   in a curl command.  (Currently is a path on netstore)
   Validate a) begins with http://netstore/, b) valid path on
   netstore, c) not userinstaller or installer-gui
   so can fail before installation process begins

3. wrapper around print statements, remove some debug prints

4. syntax change to setupadmin.sh based on Amit code review

+++ ovirt/e2einstaller.sh
@@ -2,8 +2,9 @@
 
 ###########
 #
-#  ./e2einstall.sh netstoreinstallpath node-name node-ip [True|False] 
+#  ./e2einstall.sh installerurl node-name node-ip [True|False]
 #	(if True will start up Xcalar at end)
+#   Assumes Xcalar license file XcalarLic.key exists in same dir the shell script lives in
 #
 ###########
 DIR="$(cd $(dirname ${BASH_SOURCE[0]}) && pwd)"
@@ -11,36 +12,66 @@ DIR="$(cd $(dirname ${BASH_SOURCE[0]}) && pwd)"
 # set the hostname
 echo "set hostname to $2"
 /bin/hostnamectl set-hostname $2
+rc=$?
+if [ $rc != 0 ]; then
+  exit $rc
+fi
+
 # restart so logs will see the new hostname
 echo "restart service rsyslog restart"
 service rsyslog restart
+rc=$?
+if [ $rc != 0 ]; then
+  exit $rc
+fi
 
 # install xcalar:
 # get the requested CLI installer from netstore, and kick off the installation
 # remove some yum repos first to clean things up
-echo "yo"
-curl http://netstore/$1 -o $DIR/installer.sh
+curl "$1" -o $DIR/installer.sh
+rc=$?
+if [ $rc != 0 ]; then
+  exit $rc
+fi
+
 rm -f /etc/yum.repos.d/epel*.repo /etc/yum.repos.d/mapr.repo /etc/yum.repos.d/nodesource.repo /etc/yum.repos.d/sbt.repo /etc/yum.repos.d/draios.repo /etc/yum.repos.d/ius.repo /etc/yum.repos.d/azure-cli.repo
+rc=$?
+if [ $rc != 0 ]; then
+  exit $rc
+fi
+
 chmod u+x $DIR/installer.sh
+rc=$?
+if [ $rc != 0 ]; then
+  exit $rc
+fi
 /bin/bash $DIR/installer.sh --nostart --caddy --startonboot
+rc=$?
+if [ $rc != 0 ]; then
+  exit $rc
+fi
 
 # copy in the license files
 echo 'copy lic files in to xcalar'
 cp $DIR/XcalarLic.key /etc/xcalar/XcalarLic.key
-cp $DIR/EcdsaPub.key /etc/xcalar/EcdsaPub.key
+rc=$?
+if [ $rc != 0 ]; then
+  exit $rc
+fi
 
 # generate config file.
 # if you passed only one node (single node cluster - use localhost)
 # multiple nodes passed put that node list
 echo 'generate config file via templatehelper.sh'
 chmod u+x $DIR/templatehelper.sh
+rc=$?
+if [ $rc != 0 ]; then
+  exit $rc
+fi
+
 /bin/bash $DIR/templatehelper.sh $3
-#genconfigscript='/opt/xcalar/scripts/genConfig.sh'
-#templatefile='/etc/xcalar/template.cfg'
-#fileloc='/etc/xcalar/default.cfg'
-
-#$genconfigscript $templatefile - localhost > $fileloc
-
-# now start the xcalar service
-#service xcalar start
+rc=$?
+if [ $rc != 0 ]; then
+  exit $rc
+fi
